refactor(wp): add notification for free order pending for 3 days
