+++ src/modules/addons/lknhooknotification/src/Notifications/WhatsApp/FreeOrderPendingFor3Days/lang/english.php
@@ -1,0 +1,13 @@
+<?php
+
+$lang['notification_title'] = 'Free order pending for three days';
+$lang['notification_description'] = 'Runs on DailyCronJob hook and send the notification for every client that has a free order pending for three days.';
+
+// Parameters labels
+
+$lang['order_id'] = 'Order ID';
+$lang['order_items_descrip'] = 'Order items';
+$lang['client_first_name'] = 'Client first name';
+$lang['client_full_name'] = 'Client full name';
+
+return $lang;
