+++ src/modules/addons/lknhooknotification/src/Notifications/WhatsApp/FreeOrderPendingFor3Days/lang/portuguese-br.php
@@ -1,0 +1,13 @@
+<?php
+
+$lang['notification_title'] = 'Pedido gratuito pendente há três dias';
+$lang['notification_description'] = 'Executa no pendente DailyCronJob e envia a notificação para cada cliente que tem um pedido gratuito pendente há três dias.';
+
+// Rótulos de parâmetros
+
+$lang['order_id'] = 'ID do pedido';
+$lang['order_items_descrip'] = 'Descrição itens';
+$lang['client_first_name'] = 'Nome do cliente';
+$lang['client_full_name'] = 'Nome completo do cliente';
+
+return $lang;
