+++ src/modules/addons/lknhooknotification/src/Notifications/WhatsApp/Invoice6DaysLate/lang/portuguese-pt.php
@@ -1,7 +1,7 @@
 <?php
 
 $lang['notification_title'] = 'Fatura com seis dias de atraso';
-$lang['notification_description'] = 'Executa no gancho DailyCronJob e envia a notificação para cada cliente que tem uma fatura com 6 dias de atraso.';
+$lang['notification_description'] = 'Executa no hook DailyCronJob e envia a notificação para cada cliente que tem uma fatura com 6 dias de atraso.';
 
 //Rótulos de parâmetros
